fix(rpcclient): serve WebDAV Range requests to prevent large-file corruption (#449)
* fix(rpcclient): serve WebDAV Range requests to prevent large-file corruption

The WebDAV provider declared content_range=False and ignored the HTTP Range
header, answering every GET with 200 and the entire file body. macOS Finder /
webdavfs reads large files as a series of byte ranges; receiving the whole file
in response to `Range: bytes=X-Y` makes the client write the full body at offset
X, producing a corrupted file.

Honor Range requests like asgi_webdav's own FileSystemProvider: advertise
Accept-Ranges, return 206 Partial Content with a body generator that seeks to
the requested offset and streams only the requested bytes (with 416 handling for
a failed If-Range). Also widen File.seek to c_int64 so range offsets past 2 GiB
are not truncated.


* fix(rpcclient): import ParsedType from construct.core for type-check

construct-typing 0.8.0 no longer re-exports the ParsedType TypeVar from the
top-level `construct`; it lives in construct-stubs/core.pyi. `from construct
import ParsedType` therefore fails pyright ("unknown import symbol"), turning
the type-check CI job red across all platforms. Import it from construct.core
instead. Type-checking only (guarded by TYPE_CHECKING); runtime is unaffected.

+++ src/rpcclient/rpcclient/core/subsystems/fs.py
@@ -45,7 +45,8 @@ from rpcclient.utils import cached_async_method
 
 
 if TYPE_CHECKING:
-    from construct import Construct, ParsedType
+    from construct import Construct
+    from construct.core import ParsedType
 
     from rpcclient.core.symbol import Symbol
 
@@ -148,7 +149,7 @@ class File(Allocated[ClientT_co]):
 
     async def seek(self, offset: int, whence: int) -> int:
         """Seek the remote file descriptor and return the resulting offset."""
-        err = (await self._client.symbols.lseek(self.fd, offset, whence)).c_int32
+        err = (await self._client.symbols.lseek(self.fd, offset, whence)).c_int64
         if err < 0:
             await self._client.raise_errno_exception(f"failed to lseek fd: {self.fd}")
         return err
